--- gut-to-work/app/feelings_table/page.tsx
@@ -1,11 +1,0 @@
-import { Button } from "../components/Button";
-
-export default function DashboardPage() {
-
-    return (
-        <div>
-            <Button>
-                Feelings
-            </Button>
-        </div>
-)}
