+++ gut-to-work/app/ingredients_table/page.tsx
@@ -17,7 +17,7 @@ import {
     TableHeader,
     TableRow,
 } from "../components/Table";
-import { Ingredient } from "../types";
+import { Ingredient, Ingredients } from "../types";
 
 import {
     Form,
@@ -51,6 +51,9 @@ import {
     AlertDialogDescription,
 } from "../components/AlertDialog";
 
+import { EditForm } from "../components/EditForm";
+
+
 
 
 // ----- Zod Schema for Ingredient Form ----- 
@@ -63,7 +66,7 @@ const formSchema = z.object({
 type FormData = z.infer<typeof formSchema>;
 
 export default function DashboardPage() {
-    const [ingredients, setIngredients] = useState<Ingredient[]>([]);
+    const [ingredients, setIngredients] = useState<Ingredients>([]);
     const [selectedIngredient, setSelectedIngredient] = useState<Ingredient | null>(null);
     const [open, setOpen] = useState(false);
     const [successMessage, setSuccessMessage] = useState<string | null>(null);
@@ -276,6 +279,9 @@ export default function DashboardPage() {
                     <Button className="mt-4 hover:!bg-gray-500">Create Ingredient</Button>
                 </AlertDialogTrigger>
                 <AlertDialogContent>
+                    <AlertDialogTitle>
+                        Create ingredient
+                    </AlertDialogTitle>
                     {errorMessage && (
                         <div className="text-red-600 text-sm font-medium mb-2">{errorMessage}</div>
                     )}
@@ -356,101 +362,3 @@ export default function DashboardPage() {
     );
 }
 
-type EditFormProps = {
-    ingredient: Ingredient;
-    onSubmit: (updatedIngredient: Ingredient) => void;
-    onCancel: () => void;
-    successMessage: string | null;
-    errorMessage: string | null;
-};
-function EditForm({ ingredient, onSubmit, onCancel, successMessage, errorMessage }: EditFormProps) {
-    const [name, setName] = useState(ingredient.ingredient_name);
-    const [cookingType, setCookingType] = useState(ingredient.default_cooking_type);
-    const [size, setSize] = useState(ingredient.default_portion_size);
-
-    const handleSubmit = async () => {
-        const updatedIngredient = {
-            ...ingredient,
-            ingredient_name: name,
-            default_cooking_type: cookingType,
-            default_portion_size: size,
-        };
-        try {
-            await onSubmit(updatedIngredient);
-        } catch (err) {
-            console.error("Update failed:", err);
-        }
-    };
-
-    return (
-        <div>
-            {/* Success Message */}
-            {successMessage && (
-                <div className="mb-3 p-2 rounded-md bg-green-100 text-green-700 border border-green-300 text-sm font-medium">
-                    {successMessage}
-                </div>
-            )}
-
-            {/* Error Message */}
-            {errorMessage && (
-                <div className="mb-3 p-2 rounded-md bg-red-100 text-red-700 border border-red-300 text-sm font-medium">
-                    {errorMessage}
-                </div>
-            )}
-
-            <div>
-                <label>Ingredient Name:</label>
-                <input
-                    type="text"
-                    value={name}
-                    onChange={(e) => setName(e.target.value)}
-                />
-            </div>
-
-            <div>
-                <label>Default Cooking Type:</label>
-                <DropdownMenu>
-                    <DropdownMenuTrigger asChild>
-                        <div className="w-full px-3 py-2 border border-[var(--accent)] rounded-md cursor-pointer bg-background focus:outline-none focus:ring-2 focus:ring-ring shadow-sm">
-                            {cookingType || "Select Cooking Type"}
-                        </div>
-                    </DropdownMenuTrigger>
-                    <DropdownMenuContent>
-                        <DropdownMenuSeparator />
-                        {["raw", "boiled", "deep fried", "pan fried", "baked", "infused"].map((type) => (
-                            <DropdownMenuItem key={type} onSelect={() => setCookingType(type)}>
-                                {type}
-                            </DropdownMenuItem>
-                        ))}
-                    </DropdownMenuContent>
-                </DropdownMenu>
-            </div>
-
-            <div>
-                <label>Default Size:</label>
-                <DropdownMenu>
-                    <DropdownMenuTrigger asChild>
-                        <div className="w-full px-3 py-2 border border-[var(--accent)] rounded-md cursor-pointer bg-background focus:outline-none focus:ring-2 focus:ring-ring shadow-sm">
-                            {size || "Select Portion Size"}
-                        </div>
-                    </DropdownMenuTrigger>
-                    <DropdownMenuContent>
-                        <DropdownMenuSeparator />
-                        {["small", "normal", "big"].map((sizeOption) => (
-                            <DropdownMenuItem key={sizeOption} onSelect={() => setSize(sizeOption)}>
-                                {sizeOption}
-                            </DropdownMenuItem>
-                        ))}
-                    </DropdownMenuContent>
-                </DropdownMenu>
-            </div>
-
-            <div className="flex items-center justify-end space-x-4 mt-4">
-                <AlertDialogCancel asChild>
-                    <Button variant="outline" type="button" onClick={onCancel}>Cancel</Button>
-                </AlertDialogCancel>
-                <Button onClick={handleSubmit}>Save</Button>
-            </div>
-        </div>
-    );
-}
