+++ gut-to-work/app/types.ts
@@ -7,3 +7,12 @@ export type Ingredient = {
 };
 
 export type Ingredients = Ingredient[]; // or Array<Ingredient>
+
+export type EditFormProps = {
+    ingredient: Ingredient;
+    onSubmit: (updatedIngredient: Ingredient) => void;
+    onCancel: () => void;
+    successMessage: string | null;
+    errorMessage: string | null;
+};
+
